requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
anyio==4.11.0
attrs==25.4.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.1.8
contourpy==1.3.0
cycler==0.12.1
datasets==4.4.1
dill==0.4.0
filelock==3.19.1
fonttools==4.60.1
frozenlist==1.8.0
fsspec==2025.9.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==1.1.4
idna==3.11
jinja2==3.1.6
kiwisolver==1.4.7
markupsafe==2.1.5
matplotlib==3.9.4
mir-eval==0.8.2
mpmath==1.3.0
multidict==6.7.0
multiprocess==0.70.18
numpy==2.0.2
pandas==2.3.3
pillow==11.3.0
propcache==0.4.1
pyarrow==21.0.0
pycparser==2.23
pyparsing==3.2.5
pytz==2025.2
pyyaml==6.0.3
requests==2.32.5
scipy==1.13.1
seaborn==0.13.2
shellingham==1.5.4
sniffio==1.3.1
soundfile==0.13.1
sympy==1.14.0
torchcodec==0.7.0
tqdm==4.67.1
typer-slim==0.20.0
tzdata==2025.2
urllib3==2.5.0
xxhash==3.6.0
yarl==1.22.0
